move all logic in webhook-gateway dir
